[CP-2216] Base actions merged
