+++ libs/core/settings/actions/index.ts
@@ -3,6 +3,7 @@
  * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
  */
 
+export * from "./base.action"
 export * from "./check-update-available.action"
 export * from "./load-settings.action"
 export * from "./send-diagnostic-data.action"
@@ -11,14 +12,10 @@ export * from "./set-convert.action"
 export * from "./set-diagnostic-timestamp.action"
 export * from "./set-incoming-calls.action"
 export * from "./set-incoming-messages.action"
-export * from "./set-latest-version.action"
 export * from "./set-low-battery.action"
 export * from "./set-non-standard-audio-files-conversion.action"
 export * from "./set-os-backup-location.action"
 export * from "./set-os-updates.action"
-export * from "./set-settings.action"
 export * from "./toggle-tethering.action"
 export * from "./toggle-collection-data.action"
 export * from "./toggle-privacy-policy-accepted.action"
-export * from "./toggle-application-update-available.action"
-export * from "./set-checking-for-update-action"
