+++ libs/core/settings/actions/load-settings.action.ts
@@ -7,7 +7,7 @@ import { createAsyncThunk } from "@reduxjs/toolkit"
 import isVersionGreater from "Core/utils/is-version-greater"
 import { SettingsEvent } from "Core/settings/constants"
 import { getSettings } from "Core/settings/requests"
-import { setSettings } from "Core/settings/actions/set-settings.action"
+import { setSettings } from "Core/settings/actions/base.action"
 import logger from "Core/__deprecated__/main/utils/logger"
 import { getConfiguration } from "Core/settings/requests"
 import packageInfo from "../../../../apps/mudita-center/package.json"
