--- libs/core/settings/actions/set-latest-version.action.ts
@@ -1,11 +1,0 @@
-/**
- * Copyright (c) Mudita sp. z o.o. All rights reserved.
- * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
- */
-
-import { createAction } from "@reduxjs/toolkit"
-import { SettingsEvent } from "Core/settings/constants"
-
-export const setLatestVersion = createAction<string>(
-  SettingsEvent.SetLatestVersion
-)
