+++ libs/core/settings/components/app-update-flow/app-update-flow.component.tsx
@@ -11,7 +11,7 @@ import { AppUpdateFlowTestIds } from "Core/settings/components/app-update-flow/a
 import { ModalLayers } from "Core/modals-manager/constants/modal-layers.enum"
 import { settingsStateSelector } from "Core/settings/selectors"
 import { Dispatch } from "Core/__deprecated__/renderer/store"
-import { skipAvailableUpdate } from "Core/settings/actions/skip-available-update.action"
+import { skipAvailableUpdate } from "Core/settings/actions/base.action"
 
 export const AppUpdateFlow: FunctionComponent = () => {
   const dispatch = useDispatch<Dispatch>()
